updated db connection to backend
