+++ backend/package.json
@@ -3,6 +3,8 @@
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
+    "mongodb": "^6.13.0",
+    "mongoose": "^8.9.6",
     "nodemon": "^3.1.9"
   }
 }
