--- backend/sever.js
@@ -1,29 +1,0 @@
-import express from "express";
-import cors from "cors";
-import dotenv from "dotenv";
-
-import authRoutes from "./src/routes/authRoutes.js";
-import driverRoutes from "./src/routes/driverRoutes.js";
-
-dotenv.config();
-
-const app = express();
-
-// Middleware
-app.use(cors());
-app.use(express.json());
-
-// Routes
-app.use("/api/auth", authRoutes);
-app.use("/api/drivers", driverRoutes);
-
-// Default route
-app.get("/", (req, res) => {
-    res.send("AthleteBio API is running...");
-});
-
-// Start server
-const PORT = process.env.PORT || 5000;
-app.listen(PORT, () => {
-    console.log(`Server is running on port ${PORT}`);
-});
